% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_sd.R
\name{max_sd}
\alias{max_sd}
\title{Calculate maximum Standard Deviation for a truncated variable}
\usage{
max_sd(mean, sd, n, min, max)
}
\arguments{
\item{mean}{numeric variable representing the reported mean.}

\item{sd}{numeric variable representing the reported Standard Deviation.}

\item{n}{numeric variable representing the reported sample size.}

\item{min}{numeric variable representing the variable's minimum possible/observable score.}

\item{max}{numeric variable representing the variable's maximum possible/observable score.}
}
\value{
A numeric variable representing the maximum SD.
}
\description{
Explanation to be added
}
\details{
Assumptions to be checked: Does it still work if min or max are not integers?
}
