\name{plot.Tides}
\alias{plot.Tides}
\title{Simple plot of Tides}
\description{Plot water levels, calculated high and low water, and reference level of a Tides-object}
\usage{
\method{plot}{Tides}(x,...)
}
\arguments{
  \item{x}{Tides object (e.g. the result of TidalCharacteristics())}
  \item{...}{Other arguments to be passed to plot()}
  }
\details{
  Produces a very simple plot of a Tides object, containing water levels, calculated high and low water, and reference level of a Tides-object. Useful for diagnostic purposes.
}
\author{Tom Cox <tom.cox@uantwerp.be>}
\keyword{utilities}

