\name{print.Tides}
\alias{print.Tides}
\title{Print summary information of Tides-object}
\description{Print summary information of Tides-object}
\usage{
\method{print}{Tides}(x, ...)
}
\arguments{
  \item{x}{Tides object (e.g. the result of TidalCharacteristics())}
  \item{...}{Other arguments to be passed to print()}}
\details{
  Print inundation frequency, average inundation height, average inundation time, average dry time and number of tidal cycles in time series.
}
\author{Tom Cox <tom.cox@ua.ac.be>}
\keyword{utilities}

