% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds-primitives.R
\name{sd_from_ss}
\alias{sd_from_ss}
\title{Sample SD from a sum of squared deviations}
\usage{
sd_from_ss(sum_squares, n)
}
\arguments{
\item{sum_squares}{Numeric vector, sum of squared deviations about the mean.}

\item{n}{Integer scalar, sample size.}
}
\value{
\code{sqrt(pmax(0, sum_squares) / (n - 1))}.
}
\description{
Sample SD from a sum of squared deviations
}
\keyword{internal}
