% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides_single.R
\name{tides_single}
\alias{tides_single}
\title{Calculate TIDES test for a single set of values}
\usage{
tides_single(mean, sd, n, min, max, calculate_min_sd = FALSE, verbose = TRUE)
}
\arguments{
\item{mean}{numeric variable representing the reported mean.}

\item{sd}{numeric variable representing the reported Standard Deviation.}

\item{n}{numeric variable representing the reported sample size.}

\item{min}{numeric variable representing the variable's minimum possible/observable score.}

\item{max}{numeric variable representing the variable's maximum possible/observable score.}

\item{calculate_min_sd}{logical variable representing whether a minimum SD should also be calculated. This should only be calculated if the varible is not only trucated (has a minimum and maximum possible/observable score) but also the variable is discrete/binned/granular: ie the response must be whole numbers (e.g., a 1-7 likert scale, where an indiviudal cannot provide a score of 1.5).}

\item{verbose}{logical variable representing whether the output should also contain the input values.}
}
\description{
Explanation to be added
}
