% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides_umbrella_plot.R
\name{tides_umbrella_plot}
\alias{tides_umbrella_plot}
\title{Visualize TIDES results using an umbrella plot}
\usage{
tides_umbrella_plot(
  mean,
  sd,
  n,
  min,
  max,
  calculate_min_sd = FALSE,
  show_labels = c("consistency", "values", "both", "none"),
  max_range = 10,
  color_cons = "royalblue1",
  color_incons = "red",
  point_alpha = 1,
  point_shape = 19,
  point_size = 2,
  label_force = 175,
  label_force_pull = 0.75
)
}
\arguments{
\item{mean}{Numeric. Reported mean.}

\item{sd}{Numeric. Reported standard deviation.}

\item{n}{Numeric. Reported sample size.}

\item{min}{Numeric. Minimum possible/observable score.}

\item{max}{Numeric. Maximum possible/observable score.}

\item{calculate_min_sd}{Logical. Should a minimum SD also be calculated? Only
set if to \code{TRUE} if the variable is not only truncated (it has a minimum
and maximum possible/observable score) but also discrete/binned/granular:
the response must be whole numbers (e.g., a 1-7 Likert scale, where an
individual cannot provide a score of 1.5). Default is \code{FALSE}.}

\item{show_labels}{String. One of these:
\itemize{
\item \code{"consistency"} (the default) declares whether the data are
mathematically consistent.
\item \code{"values"} displays the data in "mean (SD)" format.
\item \code{"both"} combines the two labels above.
\item \code{"none"} shows no labels.
}}

\item{max_range}{Integer. TODO: Explain \code{max_range}}

\item{color_cons}{String. Fill color of consistent data points. Default is
\code{"royalblue1"}.}

\item{color_incons}{String. Fill color of inconsistent data points. Default
is \code{"red"}.}

\item{point_alpha}{Numeric. Opacity of the data points, ranging from 0 to 1.
Default is \code{1}. Choose lower values for more transparency.}

\item{point_shape}{Numeric or string. Shape of the data points; see
https://ggplot2.tidyverse.org/articles/ggplot2-specs.html#sec:shape-spec.
Default is \code{19}.}

\item{point_size}{Numeric. Size of the data points. Default is \code{2}.}

\item{label_force}{Numeric. Force of repulsion between overlapping text
labels. Default is \code{175}.}

\item{label_force_pull}{Numeric. Force of attraction between a text label and
its corresponding data point. Default is \code{0.75}.}
}
\value{
A ggplot object: the TIDES umbrella plot.
}
\description{
TODO: Explanation to be added
}
\examples{
tides_umbrella_plot(mean = 1.1, sd = 0.5, n = 12, min = 1, max = 7)

# Assuming a discrete scale:
tides_umbrella_plot(mean = 1.1, sd = 0.5, n = 12, min = 1, max = 7, calculate_min_sd = TRUE)
}
