\name{waterlevels}
     \docType{data}
     \alias{waterlevels}
     \title{Observed water levels of the tides in the Lippenbroek Flood Control Area with controled reduced tide (FCA-CRT)}
     \description{
       This data set gives the observed water levels (in mTAW, the Belgian reference for water levels) the Lippenbroek Flood Control Area with controled reduced tide (FCA-CRT) between 2007-03-01 and 2007-04-01
     }
     \usage{waterlevels}
     \format{A data frame containing observation time, the observed water level h (mTAW) and the elevation h0 of the observation site}
     \references{
       	Cox et al (2006) \emph{Flood control areas as an opportunity to restore estuarine habitat}. Ecological engineering 28:55-36 \cr
	Maris et al (2007) \emph{Tuning the tide: creating ecological conditions for tidal marsh development in a controlled inundation area}. Hydrobiologia 588: 31-43
     }
     \keyword{datasets}

