% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_sd_bounds.R
\name{approximate_sd_bounds}
\alias{approximate_sd_bounds}
\title{Compute approximate SD bounds by filling from nearby means' SD bounds}
\usage{
approximate_sd_bounds(dat)
}
\arguments{
\item{dat}{A data frame containing the variables \code{mean}, \code{min_sd},
\code{max_sd}, \code{min}, \code{max}, \code{n}, and \code{n_items}, such
as the output of a call to \code{sd_bounds()} over a grid of means.}
}
\value{
A data frame with one row per unique mean value (and associated
metadata), containing relaxed lower and upper SD bounds (\code{min_sd},
\code{max_sd}).
}
\description{
\code{approximate_sd_bounds} applies an interpolation strategy to
relax discontinuities in SD boundary curves produced by \code{sd_bounds()}.
It combines downward and upward filled versions of the SD limits from nearby
means and returns a smoothed boundary envelope for each mean value. This is
intended to be an internal function mostly for use inside \code{plot_tides}.
}
