\name{plot.Tides}
\alias{plot.Tides}
\title{Rudimentary plot of Tides}
\description{Plot water levels, calculated high and low water, and reference level of a Tides-object}
\usage{plot.Tides(x,...)}
\arguments{
  \item{x}{Tides object (e.g. the result of TidalCharacteristics())}
  \item{...} {Other arguments to be passed to plot}
  }
\value{
  Plot
}
\author{Tom Cox <tom.cox@ua.ac.be>}
\keyword{utilities}

