% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tides.R
\name{plot_tides}
\alias{plot_tides}
\title{Calculate and plot a TIDES test for a single set of values}
\usage{
plot_tides(
  mean,
  sd,
  n,
  min,
  max,
  n_items = 1,
  digits = NULL,
  calculate_min_sd = TRUE
)
}
\arguments{
\item{mean}{numeric variable representing the reported mean.}

\item{sd}{numeric variable representing the reported Standard Deviation.}

\item{n}{numeric variable representing the reported sample size.}

\item{min}{numeric variable representing the variable's minimum possible/observable score.}

\item{max}{numeric variable representing the variable's maximum possible/observable score.}

\item{n_items}{numeric variable representing how man items were averaged over at the participant level when creating the mean. I.e., when a single item Likert scale or a sum score of a multi-item Likert scale, no prior participant level averaging occured, only averaging across participants, so items = 1. If a multi item scale and mean score across items was used, enter the number of items here.}

\item{calculate_min_sd}{logical variable representing whether a minimum SD should also be calculated. This should only be calculated if the variable is not only truncated (has a minimum and maximum possible/observable score) but also the variable is discrete/binned/granular: ie the response must be whole numbers (e.g., a 1-7 likert scale, where an indiviudal cannot provide a score of 1.5).}
}
\value{
A ggplot object: the TIDES plot and results
}
\description{
Explanation to be added
}
