% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveGaps.R
\name{RemoveGaps}
\alias{RemoveGaps}
\title{Remove list of gaps from list of intervals}
\usage{
RemoveGaps(gaps, ivals, method = c("All", "Split"))
}
\arguments{
\item{gaps}{Dataframe generated with \code{\link{gapsts}} containing gaps in waterlevel time series}

\item{ivals}{Data frame of intervals that have to be corrected for gaps (typically dry times DTS or inundation times ITs)}

\item{method}{Method to remove gaps. "All": every interval containing (part of) a gap is removed. "Split": intervals are split into smaller intervals before and after the gap}
}
