% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{tides-package}
\alias{tides-package}
\title{TIDES}
\description{
Check reported Means, SDs and Ns from truncated data (data where the measure has a min and max score) for inconsistencies.
}
\author{
\strong{Maintainer}: Ian Hussey \email{ian.hussey@unibe.ch} (\href{https://orcid.org/0000-0001-8906-7559}{ORCID})

}
