\name{Tides-package}
\alias{Tides-package}
\alias{Tides}
\alias{tides}
\docType{package}
\title{
  Characteristics of water level time series in intertidal systems.
}
\description{
  The tides package containes routines to calculate ecologicaly relevant characteristics of observed tides in intertidal systems.
}
\details{
  \tabular{ll}{
    Package: \tab Tides\cr
    Type: \tab Package\cr
    Version: \tab 2.0\cr
    License: \tab GNU Public License 2 or above\cr
  }

  
}
\author{
  Tom Cox (maintainer, <tom.cox@uantwerp.be>), 
  Lennert Schepers (contributor, <lennert.schepers@uantwerp.be>)
}
%\references{
%% to be filled
%}
\seealso{
 \code{\link{TidalCharacteristics}}: the aggregate function returning several characteristics of the tides
}

\examples{
\dontrun{
## 

}
}

\keyword{ package }
