% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{_PACKAGE}
\name{TIDES}
\alias{TIDES}
\title{TIDES}
\description{
Check reported Means, SDs and Ns from truncated data (data where the measure has a min and max score) for inconsistencies.
}
\section{Important}{

Package is in early stage development.
}

