% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides_single.R
\name{tides_single}
\alias{tides_single}
\title{Calculate TIDES test for a single set of values}
\usage{
tides_single(
  mean,
  sd,
  n,
  min,
  max,
  n_items = 1,
  digits = NULL,
  calculate_min_sd = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{mean}{numeric variable representing the reported mean.}

\item{sd}{numeric variable representing the reported Standard Deviation.}

\item{n}{numeric variable representing the reported sample size.}

\item{min}{numeric variable representing the variable's minimum possible/observable score.}

\item{max}{numeric variable representing the variable's maximum possible/observable score.}

\item{n_items}{numeric variable representing how man items were averaged over at the participant level when creating the mean. I.e., when a single item Likert scale or a sum score of a multi-item Likert scale, no prior participant level averaging occured, only averaging across participants, so items = 1. If a multi item scale and mean score across items was used, enter the number of items here.}

\item{calculate_min_sd}{logical variable representing whether a minimum SD should also be calculated. This should only be calculated if the variable is not only truncated (has a minimum and maximum possible/observable score) but also the variable is discrete/binned/granular: ie the response must be whole numbers (e.g., a 1-7 likert scale, where an indiviudal cannot provide a score of 1.5).}

\item{verbose}{logical variable representing whether the output should also contain the input values.}
}
\value{
a tibble containing the max and min SD and a summary variable \code{tides} indicating if the tested values are consistent or not. ADD NOTES ON OTHER COLUMNS RETURNED.
}
\description{
Explanation to be added. TODO: consider use of round inside the function and in the result returned.
}
\examples{
\dontrun{
tides_single(mean = 3.10, sd = 0.80, n = 1100, min = 1, max = 7, n_items = 1, digits = 2) |>
  plot_tides_single()
}

}
