% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides_single.R
\name{tides_single}
\alias{tides_single}
\title{Calculate TIDES test for a single set of values}
\usage{
tides_single(mean, sd, n, min, max, calculate_min_sd = FALSE, verbose = TRUE)
}
\arguments{
\item{mean}{Numeric. Reported mean.}

\item{sd}{Numeric. Reported standard deviation.}

\item{n}{Numeric. Reported sample size.}

\item{min}{Numeric. Minimum possible/observable score.}

\item{max}{Numeric. Maximum possible/observable score.}

\item{calculate_min_sd}{Logical. Should a minimum SD also be calculated? Only
set if to \code{TRUE} if the variable is not only truncated (it has a minimum
and maximum possible/observable score) but also discrete/binned/granular:
the response must be whole numbers (e.g., a 1-7 Likert scale, where an
individual cannot provide a score of 1.5). Default is \code{FALSE}.}

\item{verbose}{Logical. Should the output also contain the input values?
Default is \code{TRUE}.}
}
\value{
A tibble containing the max and min SD and a summary variable
\code{consistency} indicating if the tested values are consistent or not.
}
\description{
Explanation to be added
}
\examples{
tides_single(mean = 1.1, sd = 0.5, n = 12, min = 1, max = 7)
}
