% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tides.R
\name{umbrella_plot}
\alias{umbrella_plot}
\title{Calculate and plot a TIDES test for a single set of values}
\usage{
umbrella_plot(
  mean,
  sd,
  n,
  min,
  max,
  calculate_min_sd = FALSE,
  show_labels = c("consistency", "values", "both", "none"),
  color_cons = "royalblue1",
  color_incons = "red",
  point_alpha = 1,
  label_force = 175,
  label_force_pull = 0.75
)
}
\arguments{
\item{mean}{Numeric. Reported mean.}

\item{sd}{Numeric. Reported standard deviation.}

\item{n}{Numeric. Reported sample size.}

\item{min}{Numeric. Minimum possible/observable score.}

\item{max}{Numeric. Maximum possible/observable score.}

\item{calculate_min_sd}{Logical. Should a minimum SD also be calculated? Only
set if to \code{TRUE} if the variable is not only truncated (it has a minimum
and maximum possible/observable score) but also discrete/binned/granular:
the response must be whole numbers (e.g., a 1-7 Likert scale, where an
individual cannot provide a score of 1.5). Default is \code{FALSE}.}

\item{show_labels}{String. One of these: \code{"consistency"} says whether the
data are numerically consistent; \code{"values"} displays them in "mean (SD)"
format; \code{"both"} combines the two labels, and \code{"none"} shows neither.}

\item{point_alpha}{Numeric. Opacity of the data points. Default is \code{1}.}

\item{label_force}{Numeric. Force of repulsion between overlapping text
labels. Default is \code{175}.}

\item{label_force_pull}{Numeric. Force of attraction between a text label and
its corresponding data point. Default is \code{0.75}.}
}
\value{
A ggplot object: the TIDES umbrella plot and results.
}
\description{
Explanation to be added
}
\examples{
umbrella_plot(mean = 1.1, sd = 0.5, n = 12, min = 1, max = 7)
}
